refactor (pilot-app/extension): transaction deletion (#1433)
## Description:

Closes #1347 

This PR updates the transaction deletion behavior in response to issue
[#1347](https://github.com/gnosis/zodiac-pilot/issues/1347). Previously,
dispatching REMOVE_TRANSACTION would remove the selected transaction and
all subsequent ones. While this was originally intentional, the current
requirement is to make this behavior explicit and more predictable.

This change introduces a new action, `CLEAR_TRANSACTIONS`, to handle the
removal of a transaction and all later ones, while keeping
`REMOVE_TRANSACTION` scoped to only the transaction with the specified
ID.