fix(extension): link to requested permissions no longer goes to Safe{Wallet} (#1442)
Opening the roles app directly in the Safe doesn't work so well. There
are issues with the authentication mechanism. So I slightly adjusted the
flow:

Instead we now open the stand-alone roles app. I've adjusted the roles
app so that upon applying it will then redirect to Safe wallet so the
user can submit the update transaction there.

---------

Co-authored-by: Jan-Felix <jfschwarz@users.noreply.github.com>