chore: transaction queues (#1590)
This PR is a huge overhaul of how we handle transactions inside the
extension.

## TL;DR
All actions are derived from the transaction's state. It now has a
`pending` and `executed` queue. Additionally, I've added the fields
`rollback` and `refresh` to indicate that either a transaction needs to
be rolled back or that everything should be done again on a fresh fork.
Also, the `ForkProvider` doesn't call handlers directly anymore but
emits an event when a transaction is supposed to be submitted and then
waits for another event to signal that the transaction has been handled.
This decouples the provider completely from our internal logic.

I've also introduced a bunch of new actions for the transaction state
reducer. This was needed as all features now need to be represented by
the state itself. On the other side, there are a couple of focused hooks
that will react differently to state updates. I've also added a bunch of
tests (but probably am still missing some)