feat(pilot-extension): Support Cow TWAPs (#1898)
So far CowSwap disables its TWAP feature when the Safe is connected via
Pilot. Some digging into their frontend code revealed the possibility to
make it work using [EIP-5792](https://eips.ethereum.org/EIPS/eip-5792).

Apparently 5792 is expected to become more popular with the availability
of 7702. So it definitely makes sense to build out Pilot in this
direction. This PR only delivers an incomplete 5792 implementation, but
it's at least a start and enough to unlock Cow TWAPs. 🎉

---------

Co-authored-by: Jan-Felix <jfschwarz@users.noreply.github.com>