feat(companion-app): extension onboarding (#2069)
We now show clearly if the Zodiac Pilot extension is not yet installed
and offer a link to the Chrome Web Store.

This is getting more important as Zodiac OS becomes the primary
entrypoint and the extension acts more to progressively enhance the user
experience for Zodiac OS users. Also in the context of Railgun Connect,
first-time users are now guided clearly through the initial set up flow.

---------

Co-authored-by: Jan-Felix <jfschwarz@users.noreply.github.com>