chore(zodiac-os): reduce number of requests to debank (#2199)
Part of #2198 

We made a request to debank on almost every page load because we were
getting a list of chains. We then only used that list to render chain
icons in some places. I've extracted the icon part into a separate
endpoint and also set a cache header on it. So this endpoint should only
re-validate every week.

This alone should reduce the number of requests we're making drastically
since the chain list request is responsible to 83% of our token usage.